+++ packages/Server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mj-biz-apps/accounting-server
 
+## 0.12.0
+
+### Patch Changes
+
+- Updated dependencies [ba62339]
+- Updated dependencies [541595e]
+- Updated dependencies [a64b1e0]
+- Updated dependencies [9659501]
+  - @mj-biz-apps/accounting-actions@0.12.0
+  - @mj-biz-apps/accounting-core-entities-server@0.12.0
+  - @mj-biz-apps/accounting-entities@0.12.0
+
 ## 0.11.0
 
 ### Patch Changes